fix godoc parsing

+++ internal/model/bids.go
@@ -6,14 +6,14 @@ import (
 
 // Bid ...
 type Bid struct {
-	ID             int       `db:"id"`
-	FileID         int       `db:"file_id"`
+	ID             int64     `db:"id"`
+	FileID         int64     `db:"file_id"`
 	WorkflowStatus int       `db:"workflow_status"`
 	Code           string    `db:"code"`
-	District       int       `db:"district"`
+	District       int64     `db:"district"`
 	PassType       int       `db:"type"`
 	CreatedAt      time.Time `db:"created_at"`
-	CreatedBy      int       `db:"created_by"`
-	UserID         int       `db:"user_id"`
+	CreatedBy      int64     `db:"created_by"`
+	UserID         int64     `db:"user_id"`
 	Source         string    `db:"source"`
 }
